changed functionality to draw a quiver plot
This update should generate the required matrices for a 3d quiver plot and then plot the vectors in a line along x up to 256 points long.

I'd recommend figuring out some way to rotate the device at a fixed rate if you wanted to plot the field vectors more spatially accurately.

One other suggestion, to address the latency when plotting, is to instead of calling the quiver/scatter function for each frame, update the data objects for your plots. This avoids redrawing the entire axis. I *think* this functionality exists in the 3d plots in MPL, as I know it exists in the 2d versions.